+++ source/windowing/StarGuiContext.hpp
@@ -106,7 +106,7 @@ public:
   void setFontProcessingDirectives(String const& directives);
   void setFont(String const& font);
   void setDefaultFont();
-  TextStyle& setTextStyle(TextStyle const& textStyle, int pixelRatio);
+  TextStyle& setTextStyle(TextStyle const& textStyle, float pixelRatio);
   TextStyle& setTextStyle(TextStyle const& textStyle);
   void clearTextStyle();
